+++ src/Maikebing.Data.Taos/TaosException.cs
@@ -41,7 +41,7 @@ namespace Maikebing.Data.Taos
             te.Data.Add("commandText", _commandText);
             throw te;
         }
-        public static void ThrowExceptionForRC(long _taos)
+        public static void ThrowExceptionForRC(IntPtr _taos)
         {
             var te = new TaosException(new TaosErrorResult() { Code = TDengine.ErrorNo(_taos), Error = TDengine.Error(_taos) });
             throw te;
